fix: teleport page - balance not updated
